+++ src/Slider/SlideTrack/SlideTrack.types.ts
@@ -1,0 +1,3 @@
+export interface SlideTrackPropsType {
+  handleTrackClick: (e: React.MouseEvent<HTMLSpanElement> | MouseEvent) => void;
+}